requirements.txt
requests>=2.25.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0
sqlalchemy>=1.4.0
pymysql>=1.0.0
python-dotenv>=0.19.0